symbols/nokia_vndr/su-8w: make a variant work that never worked
This variant, "us_nodeadkeys", included another layout, "us_intl",
that does not exist, and did not exist when this file was added.
The intended inclusion seems to have been simply "us".

Also, correct the changed key from <AC10> to <AC11>, otherwise
the layout would be without a semicolon/colon key.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>