rules/compat: remove the incomplete and faulty rules for Dvorak layouts
They are incomplete because, for example, 'setxkbmap dvorak it' will
give you the American Dvorak layout, not the Italian one.  They are
faulty because, for example, 'setxkbmap dvorak ca' fails (because the
rules expect the layout to be called "dvorak", not "fr-dvorak").

These rules may have been useful during a transition period of a few
years, but now they are just anomalies that distract from the basic
format of 'setxkbmap country-code variant-name'.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>