delete two space-bar options that behave identically to two others
With an explicit key type, specifying 'NoSymbol' for an upper level
is exactly the same as leaving the level empty.  The latter is what
the options without the trailing 's' do.

(The intent was probably to use VoidSymbol instead, but as no one has
complained that these options don't actually clear the fourth level,
and I can't see any definitions of `<SPCE>` where it would make sense
to clear the hard-to-type fourth level, just drop these options.)

Also trim two unneeded comments that are likely to become outdated.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>