symbols/lt: condense comments, and use an include instead of duplication
As the "LEKPa" is almost identical to the "LEKP" layout, it was wasteful
to fully define the keymap explicitly instead of using an 'include'.

Furthermore, avoid an include of 'eurosign(e)' when the `AD03` key
(or neighboring key) gets defined anyway.  And reshuffle a few keys
into their usual order.