symbols/ir: use the correct symbol name for the € sign
Commit c8ca0256d6 from seventeen years ago introduced (among many
other things), the strange code 0x13a4 on the third level of AD03.
According to the Iranian national standard for Persian keyboard
layout, ISIRI 9147 [1], there should be a euro sign (€) on that
third level.  Maybe at some time 0x13a4 encoded €, but it doesn't
do so nowadays.  So... use the correct symbol name instead.

[1] http://persian-computing.org/archives/ISIRI/ISIRI-9147.pdf

This fixes issue #347.

With-help-from: Alborz Jafari

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>