rules: rename an option, and compose its effect from two others
Making the CapsLock key into a Control key, and the left Control key
into a group toggle, can be achieved by combining two existing options
-- they just need to be combined in the correct order: the group toggle
should override the second half of `swapcaps`.

The option is also renamed, to more clearly express its result.

An alias is added for compatibilty.  (An alias for another option
is also added, as commit 7124dd5127 put it in the wrong place.)

This fixes issue #349.