add guillemets to the main Esperanto layout
Guillemets are regularly used as quotation marks in Esperanto texts
online, and are accepted as a form of alternative punctuation, but
were not included in the official Esperanto layout created in 2006.

This commit adds double guillemets to level 3 of keys AB08 and AB09,
and single guillemets to level 4 (matching with < and > on level 2).

This fixes issue #352.

Requested-by: J.M.C. Williams

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>